pasta de upload

+++ application/controllers/Equipe.php
@@ -16,7 +16,7 @@ class Equipe extends CI_Controller {
         //chama o metodo que faz a validção de login de usuario
         $this->load->model('Usuario_model');
         $this->Usuario_model->verificaLogin();
-        $this->load->model('Equipe_model', 'cm');
+        $this->load->model('Equipe_model');
     }
 
     public function index() {
@@ -25,7 +25,7 @@ class Equipe extends CI_Controller {
     }
 
     public function listar() {
-        $data['equipes'] = $this->cm->getALL();
+        $data['equipes'] = $this->Equipe_model->getALL();
         $this->load->view('Header');
         $this->load->view('ListaEquipes', $data);
         $this->load->view('Footer');
@@ -39,7 +39,7 @@ class Equipe extends CI_Controller {
             $this->load->view('FormEquipe');
             $this->load->view('Footer');
         } else {
-            $this->load->model('Equipe_model');
+          
             $data = array(
                 'nome' => $this->input->post('nome'),
             );
@@ -52,12 +52,12 @@ class Equipe extends CI_Controller {
             if (!$this->upload->do_upload('imagem')) {
                 $error = $this->upload->display_errors();
                 //cria uma sessão com o error e o redireciona
-                $this->session->set_flashdata('mensagem', '<div class=alert alert success>Equipe foi cadastrada com sucesso</div>');
+                $this->session->set_flashdata('mensagem');
                 redirect('Equipe/listar'); //Se der certo manda para a lista 
                 exit();
             } else {
                 //pega o nome do arquivo que foi enviado e adiciona no array $data que
-                $data['imagem'] = $this->uploads->data('file_name');
+                $data['imagem'] = $this->upload->data('file_name');
             }
             if ($this->Equipe_model->insert($data)) {
                 $this->session->set_flashdata('mensagem', '<div class="alert alert sucees>Sucesso</div>"');
@@ -71,11 +71,11 @@ class Equipe extends CI_Controller {
 
     public function alterar($id) {
         if ($id > 0) {
-            $this->load->model('Equipe_model');
+
             $this->form_validation->set_rules('nome', 'nome', 'required');
 
             if ($this->form_validation->run() == false) {
-                $data['equipe'] = $this->Equipe_model->getONE($id);
+                $data['equipe'] = $this->Equipe_model->getOne($id);
 
                 $this->load->view('Header');
                 $this->load->view('FormEquipe', $data);
@@ -83,6 +83,7 @@ class Equipe extends CI_Controller {
             } else {
                 $data = array(
                     'nome' => $this->input->post('nome'),
+                    'imagem' => $this->input->post('imagem'),
                 );
 
                 $config['upload_path'] = './uploads/';
@@ -91,7 +92,8 @@ class Equipe extends CI_Controller {
                 $config['max_height'] = 768;
                 $config['encrypt_name'] = true;
                 $this->load->library('upload', $config);
-                if (!$this->upload->do_upload('userfile')) {
+               
+                if (!$this->upload->do_upload('imagem')) {
                     $error = $this->upload->display_errors();
                     //cria uma sessão com o error e o redireciona
                     $this->session->set_flashdata('mensagem', '<div class=alert alert success>Imagem foi cadastrada com sucesso</div>');
@@ -99,10 +101,11 @@ class Equipe extends CI_Controller {
                     exit();
                 } else {
                     //pega o nome do arquivo que foi enviado e adiciona no array $data que
-                    $data['imagem'] = $this->uploads->data('file_name');
+                    $data['imagem'] = $this->upload->data('file_name');
                 }
 
                 if ($this->Equipe_model->update($id, $data)) {
+                    unlink('uploads/' . $equipe->imagem);
                     redirect('Equipe/listar');
                 } else {
                     redirect('Equipe/alterar' . $id);
@@ -115,11 +118,14 @@ class Equipe extends CI_Controller {
 
     public function deletar($id) {
         if ($id > 0) {
-            $this->load->model('Equipe_model');
-            if ($this->Equipe_model->delete($id)) {
-                $this->session->set_flashdata('mensagem', 'Equipe deletada com sucesso!');
-            } else {
-                $this->session->set_flashdata('mensagem', 'Falha ao deletar Equipe...');
+            $equipe = $this->Equipe_model->getOne($id);
+            if ($equipe) {
+                if ($this->Equipe_model->delete($id)) {
+                    unlink('uploads/' . $equipe->imagem);
+                    $this->session->set_flashdata('mensagem', 'Equipe deletada com sucesso!');
+                } else {
+                    $this->session->set_flashdata('mensagem', 'Falha ao deletar Equipe...');
+                }
             }
         }
         redirect('Equipe/listar');
